backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
requests
beautifulsoup4
python-dotenv
cohere
